rename, fix svgr
